Check before displaying
Because the image request is being done in the background, we need to
check that the table row that originally requested it is still in view
before displaying it. Else, by the time the image comes back, the user
could have already scrolled to a different cell and we don't want to
update that one wrongly.

Either way, the image is cached for future use so that the next time,
the image for this row is requested, we can just load it from cache.